backend/requirements.txt
python-dotenv
django-ninja
django-ninja-extra
django-ninja-jwt
pydantic[email]